requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
pydantic>=2.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
torch
torchvision
torchaudio
transformers
accelerate
sentencepiece
protobuf
pillow
mlx>=0.20.0
mlx-lm>=0.20.0
